[#60] Remove extend_auth
